token check changes

+++ data/Job/save_job.py
@@ -19,21 +19,24 @@ def save_job(request):
     token = data.get('token')
     user_id,registered_by,email = decode_token(token)
     print(user_id, registered_by,email)
-    # Check if the record already exists
-    existing_record = session.query(SavedJob).filter_by(user_id=user_id, job_id=job_id).first()
-    
-    if existing_record:
-      # If the record already exists, return an error
-      return message.response('Error', 'alreadySavedJobError')
+    if user_id is not None:
+      # Check if the record already exists
+      existing_record = session.query(SavedJob).filter_by(user_id=user_id, job_id=job_id).first()
+      
+      if existing_record:
+        # If the record already exists, return an error
+        return message.response('Error', 'alreadySavedJobError')
+      else:
+        save_job_instance = SavedJob(user_id=user_id, job_id=job_id)
+        session.add(save_job_instance)
+        session.commit()
+        session.close()
+      if save_job_instance:  # Checking if response_data is not empty
+        return message.response('Success', 'savedJob')
+      else:
+        return message.response('Error', 'savedJobError')
     else:
-      save_job_instance = SavedJob(user_id=user_id, job_id=job_id)
-      session.add(save_job_instance)
-      session.commit()
-      session.close()
-    if save_job_instance:  # Checking if response_data is not empty
-      return message.response('Success', 'savedJob')
-    else:
-      return message.response('Error', 'savedJobError')
+      return message.response('Error', 'tokenError')
   except Exception as e:
     return message.tryExceptError(str(e))
   
@@ -46,15 +49,18 @@ def delete_save_job(request):
     token = data.get('token')
     user_id,registered_by,email = decode_token(token)
     print(user_id, registered_by,email)
-    saved_job_instance = session.query(SavedJob).filter_by(user_id=user_id, job_id=job_id).first()
-    if saved_job_instance:
-      session.delete(saved_job_instance)
-      session.commit()
-      session.close()
-      return message.response('Success', 'savedUnJob',)
+    if user_id is not None:
+      saved_job_instance = session.query(SavedJob).filter_by(user_id=user_id, job_id=job_id).first()
+      if saved_job_instance:
+        session.delete(saved_job_instance)
+        session.commit()
+        session.close()
+        return message.response('Success', 'savedUnJob',)
+      else:
+        session.close()
+        return message.response('Error', 'savedJobError')
     else:
-      session.close()
-      return message.response('Error', 'savedJobError')
+      return message.response('Error', 'tokenError')
   except Exception as e:
     return message.tryExceptError(str(e))
   
@@ -66,22 +72,25 @@ def get_all_saved_jobs(request):
     token = data.get('token')
     user_id, registered_by, email = decode_token(token)
     print(user_id, registered_by, email)
-    saved_job_ids = session.query(SavedJob.job_id).filter_by(user_id=user_id).all()
-    # Extract job IDs from the result and convert them into a list
-    job_ids_list = [job_id[0] for job_id in saved_job_ids]
-    print(job_ids_list)
-    response_data = []
-    set_data_id = set()
-    for job_id in job_ids_list:
-      if job_id in set_data_id:
-        continue
-      set_data_id.add(job_id)
-      job_result = job_details_query.job_result(job_id, set_data_id)
-      job_result_dict = json.loads(job_result)  # Convert search_result to a Python dictionary
-      response_data.append(job_result_dict)  # Append the job data inside the loop
-    if response_data:
-      return message.response1('Success', 'userApplyJob', response_data)
+    if user_id is not None:
+      saved_job_ids = session.query(SavedJob.job_id).filter_by(user_id=user_id).all()
+      # Extract job IDs from the result and convert them into a list
+      job_ids_list = [job_id[0] for job_id in saved_job_ids]
+      print(job_ids_list)
+      response_data = []
+      set_data_id = set()
+      for job_id in job_ids_list:
+        if job_id in set_data_id:
+          continue
+        set_data_id.add(job_id)
+        job_result = job_details_query.job_result(job_id, set_data_id)
+        job_result_dict = json.loads(job_result)  # Convert search_result to a Python dictionary
+        response_data.append(job_result_dict)  # Append the job data inside the loop
+      if response_data:
+        return message.response1('Success', 'userApplyJob', response_data)
+      else:
+        return message.response1('Error', 'searchJobError', data={})
     else:
-      return message.response1('Error', 'searchJobError', data={})
+      return message.response('Error', 'tokenError')
   except Exception as e:
     return message.tryExceptError(str(e))
